requirements.txt
boto3
azure-identity
azure-mgmt-costmanagement
google-cloud-bigquery
google-auth
pymongo
schedule
python-dotenv
